Patrick's tests moved to renamed "screens" folder
