+++ app/src/androidTest/java/com/android/partagix/authentication/BootTest.kt
@@ -4,7 +4,7 @@ import androidx.compose.ui.test.*
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.android.partagix.model.auth.Authentication
-import com.android.partagix.screen.BootScreen
+import com.android.partagix.screens.BootScreen
 import com.android.partagix.ui.navigation.NavigationActions
 import com.android.partagix.ui.navigation.Route
 import com.android.partagix.ui.screens.BootScreen
